Render world through raycasting (in a rough way)
Finally renders world using raycast.
Code is rough as hell and DDA implementation is taken from the internet (and
i dont like that) so i'm planning to rewrite the DDA function.
Also this code is not modular at all and does not allow for much as of
now...

Update REAMDE.md